Move all admin presenters to their own subfolder
