+++ src/professions/admin/interfaces/index-template.interface.ts
@@ -1,7 +1,7 @@
 import { CheckboxItems } from 'src/common/interfaces/checkbox-items.interface';
 import { Table } from '../../../common/interfaces/table';
 import { RegulationType } from '../../profession-version.entity';
-import { ProfessionsPresenterView } from '../professions.presenter';
+import { ProfessionsPresenterView } from '../presenters/professions.presenter';
 
 export interface IndexTemplate {
   view: ProfessionsPresenterView;
